chores: made total points work well

+++ members/views.py
@@ -4,7 +4,8 @@ from django.urls import reverse
 from forms import MemForm
 from lottery.functions.sum_points import sum_points
 from lottery.models import Point
-from members.models import Member
+from members.models import Member, Level
+
 
 # Create your views here.
 def index(request):
@@ -14,22 +15,20 @@ def index(request):
     }
     return render(request, "index.html", context)
 
-
 def create_member(request):
     # POST라면 입력한 내용을 form을 이용하여 데이터베이스에 저장
     if request.method == 'POST' or request.method == 'FILES':
-        group = str(request.POST.get('group_time_mode')) + str(request.POST.get('group_level_mode'))
         form = MemForm(request.POST, request.FILES)
 
         # 유효성 검사
         if form.is_valid():
-            print("group test: " + form.cleaned_data['group'])
-
+            form.cleaned_data['total_points'] = request.POST.get('total_points')
+            form.save()
             # form.cleaned_data['group'] = group
-            record = form.save(commit=False)
-            record.group = group
-            record.save()
-            print("group: " + form.cleaned_data['group'])
+            # record = form.save(commit=False)
+            # record.group = level
+            # record.save()
+            # print("group: " + form.cleaned_data['group'])
 
             return redirect('lottery:index')
 
@@ -40,10 +39,9 @@ def create_member(request):
     return render(request, 'create_member_form.html', {'form': form})
 
 
-
 def detail(request, member_id):
     member = get_object_or_404(Member, pk=member_id)
-    point_list = Point.objects.order_by('-date')
+    point_list = Point.objects.order_by('-date').filter(member=member)
 
     total_points = sum_points(member_id)
     num_chances = total_points / 30
@@ -61,9 +59,11 @@ def detail(request, member_id):
     return render(request, "detail.html", context)
 
 
+def delete_member(request, member_id):
 
-def delete_member(member_id):
-    member = get_object_or_404(Member, pk=member_id)
-    member.delete()
+    if request.method == 'POST':
+        member = get_object_or_404(Member, pk=member_id)
+        member.delete()
+
     return redirect(reverse('members:index'))
